api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.30.0
google-cloud-firestore>=2.20.0
pydantic>=2.9.0
python-multipart>=0.0.12
